Fixed firebase monitoring
Need to make sure entries don't already exist
